Перенос изображений, редиректы, счетчик редиректов

+++ protected/models/Redirects.php
@@ -16,13 +16,23 @@ class Redirects extends CActiveRecord
     
     public function beforeSave()
     {
-        $criteria = new CDbCriteria();
-        $criteria->condition = '`from` = :from or `to` = :to or `from` = :to or `to` = :from';
-        $criteria->params = [
-            'from'=> $this->from, 
-            'to'=>  $this->to
-        ];
-        self::model()->deleteAll($criteria);
+        if ($this->isNewRecord) {
+            $criteria = new CDbCriteria();
+            $criteria->condition = '`from` = :from or `to` = :to or `from` = :to or `to` = :from';
+            $criteria->params = [
+                'from'=> $this->from, 
+                'to'=>  $this->to
+            ];
+            self::model()->deleteAll($criteria);
+        }
         return parent::beforeSave();
+    }
+    
+    public function updateCounter()
+    {
+        if (!$this->isNewRecord) {
+            $this->counter = $this->counter + 1;
+            $this->save();
+        }
     }
 }
